+++ src/transport/smol/mod.rs
@@ -37,7 +37,6 @@ mod tests {
         TransportConfig {
             buffer_config: BufferConfig {
                 recv_buffer_size: 65,
-                send_buffer_size: 64,
                 max_buffer_size: 64,
             },
             ..TransportConfig::default()