+++ src/transport/tokio/serial.rs
@@ -216,7 +216,6 @@ mod tests {
             .if_clear_on_connect(false)
             .buffer_config(BufferConfig {
                 recv_buffer_size: 65,
-                send_buffer_size: 64,
                 max_buffer_size: 64,
             });